refactor: Refactor project structure and dependencies
- Renamed files and updated package declarations for `Cluster.kt`, `RpcInterface.kt`, `Transaction.kt`, `Signer.kt`, `Message.kt`
- Added dependencies for `solana_eddsa`, `solana_keypair`, `solana_public_keys`, `solana_interfaces`, `base58`, `com.ditchoom:buffer`, `com.diglol.crypto:crypto`
- Deleted `Umi.kt`, `TransactionBuilder.kt`
- Added new files `solana_keypair/build.gradle.kts`, `TransactionBuilder.kt`, `solana_interfaces/src/commonMain/kotlin/com/metaplex/solana_interfaces/TransactionBuilder.kt`, `Keypair.kt`, `SolanaEddsa.kt`, `PublicKeyTests.kt`, `solana_public_keys/build.gradle.kts`, `PublicKey.kt`, `SolanaKeypair.kt`, `UmiTransactionBuilder.kt`, `EddsaInterface.kt`, `TransactionInterface.kt`, `UmiTransaction.kt`, `SolanaKeypair.kt`, `SolanaEddsa.kt`, `solana_interfaces/build.gradle.kts`, `UmiMessage.kt`, `MessageInterface.kt`
- Renamed and modified various files and imports
- Added new interfaces and classes in `solana_interfaces/src/commonMain/kotlin/com/metaplex/solana_interfaces`
