+++ settings.gradle.kts
@@ -16,5 +16,8 @@ dependencyResolutionManagement {
 
 rootProject.name = "umi"
 include(":umi")
-include(":umi_public_keys")
+include(":solana_public_keys")
 include(":base58")
+include(":solana_eddsa")
+include(":solana_interfaces")
+include(":solana_keypair")
