+++ solana_interfaces/src/commonMain/kotlin/com/metaplex/solana_interfaces/EddsaInterface.kt
@@ -1,0 +1,21 @@
+package com.metaplex.solana_interfaces
+
+import com.metaplex.solana_public_keys.Pda
+import com.metaplex.solana_public_keys.PublicKey
+
+interface EddsaInterface {
+    /** Generates a new keypair. */
+    suspend fun generateKeypair(): Keypair
+    /** Restores a keypair from a secret key. */
+    suspend fun createKeypairFromSecretKey(secretKey: ByteArray): Keypair
+    /** Restores a keypair from a seed. */
+    suspend fun createKeypairFromSeed(seed: ByteArray): Keypair
+    /** Whether the given public key is on the EdDSA elliptic curve. */
+    suspend fun isOnCurve(publicKey: PublicKey): Boolean
+    /** Finds a Program-Derived Address from the given programId and seeds. */
+    suspend fun findPda(programId: PublicKey, seeds: Array<ByteArray>): Pda
+    /** Signs a message with the given keypair. */
+    suspend fun sign(message: ByteArray, keypair: Keypair): ByteArray
+    /** Verifies a signature for a message with the given public key. */
+    suspend fun verify(message: ByteArray, signature: ByteArray, publicKey: PublicKey): Boolean
+}
