+++ solana_public_keys/build.gradle.kts
